Added jumping and improved tiles
Fixed the collisions on the tiles by clicking in the top left of the sprite editor and changing the custom physics collisions. This allowed me to make some of the colliders for the tiles diagonal, etcetera. The jump mechanic, could be polished slightly further for the jetpack and the tiles could be made better by tweaking the applied physics material.